Objective3Lab1

+++ Labs/Objective3Lab1.java
@@ -2,7 +2,7 @@ public class Objective3Lab1 {
   public static void main(String[] args) {
     double num1, num2, num3, average;
 
-    Integer num1 = 4;
+    double num1 = 4;
     System.out.println(num1);
     double num2 = 10;
     System.out.println(num2);
